Change the way recipients are counted.
In a case of a mail with multiple recipient, if the nth recipient trigger
the rate limit, discard the previous recipient from the database : indeed
no mail was sent as postfix need to accept all recipient or none.

Futhermore, migrate from using the RCPT state for counting the recipient to
the DATA state + recipient_count parameter.

For this to work, postfix config needs to be edited check_policy_service
needs to be moved from smtpd_recipient_restrictions to smtpd_data_restrictions
and the settings count_mode needs to be set to 1 in policyd-rate-limit.yaml.

The default value of count_mode is 0 is not defined to avoir breaking
existing setups upon upgrade but is marked as deprecated.